Fixes: https://github.com/eclipse/openj9/issues/7332
Doc issue: https://github.com/eclipse/openj9-docs/issues/413

Ports:
ibmruntimes/openj9-openjdk-jdk11#226
ibmruntimes/openj9-openjdk-jdk13#25
ibmruntimes/openj9-openjdk-jdk#149

**Summary:**
- The assumption that the ludcl will not change during a call to readObject/readUnshared can only be made if user code is not invoked. This change triggers a refresh of the cachedLudcl when a user method is invoked.
- Re-enable ludcl caching by default (com.ibm.enableClassCaching)

**Performance assessment:** (pending)
- Daytrader7: as a general benchmark for OpenJ9
- second benchmark that to more clearly show benchmark benefits: I am working with the perf team to get results for this. From previous documentation the patches in ObjectInputStream were critical to its performance

Signed-off-by: Theresa Mammarella <Theresa.T.Mammarella@ibm.com>