@DanHeidinga 2019-10-25
> pending: performance impact assessment

+1.  Can you expand a bit on what will be tested to validate the performance of this change?