@theresa-m 2019-10-25
> > pending: performance impact assessment
> 
> +1. Can you expand a bit on what will be tested to validate the performance of this change?

- Daytrader7: as a general benchmark for OpenJ9
- SPECjbb2015: I am working with the perf team to get results for this. From previous documentation the patches in ObjectInputStream were critical to its performance