@theresa-m 2019-10-30
These are my performance results. The "without changes" builds have the optimization enabled by default.

Daytrader7:

  | Startup mean | Throughput mean | Throughput Warm mean
-- | -- | -- | --
With Changes | 4456 | 5578 | 5797
Without Changes | 4512 | 5656 | 5570
% difference | 1.25 | 1.40 | -3.92

SpecJbb2015 ppc:

  | Max-jops mean | Crit-jops mean
-- | -- | --
With changes | 136650 | 46173
Without changes | 140096 | 47778
% difference | 2.52 | 3.47