@zl-wang 2019-10-30
For consistency, only "Startup" is lower meaning better; all "Throughput" is lower meaning worse (regressed). Please mark the + and - % accordingly.