@piyush286 2019-10-31
![image](https://user-images.githubusercontent.com/10422081/67976368-577a5780-fbec-11e9-8b83-7e916c1f96e6.png)

For more detailed results, please see this spreadsheet: https://ibm.box.com/s/sztz4lmb3urxvkjujqnn5wzrjv0mu1eu

There was one outlier in each of the sets that showed high CI. In the spreadsheet, I've marked the outliers in red. As you can in the above screenshot, I've added a column for `Without Outlier`, reflecting a more accurate picture. 

@theresa-m Hopefully, I haven't made any calculation mistakes, but feel free to double check. Btw it might be good to look at CIs for SPECjbb2015 runs as well.