@theresa-m 2019-11-04
Here are more results with the second benchmark, including with the ludcl optimization disabled:

  | Max-jops mean | Crit-jops mean
-- | -- | --
1 With Changes + LUDCL Enabled | 136650 | 46173
2 Without Changes + LUDCL Enabled | 140096 | 47778
3 LUDCL Disabled | 131298 | 45036
1 vs 2 (%) | 97.54 | 96.64
1 vs 3 (%) | 104.08 | 102.53
2 vs 3 (%) | 106.70 | 106.09

[ludcl_benchmark2.xlsx](https://github.com/ibmruntimes/openj9-openjdk-jdk8/files/3804377/ludcl_benchmark2.xlsx)