@DanHeidinga 2019-11-04
Summarizing the results so far:
* Theresa's fix makes the LUDCL optimization correct.
* It is an improvement over having the LUDCL optimization disabled.
* It's not quite as good as the broken version but being correct is better :)

This patch is worth delivering to restore most of the optimization.

Anyone want to weight in the above summary?
@theresa-m @vijaysun-omr @zl-wang @piyush286