@DanHeidinga 2019-11-05
@theresa-m can the WIP be removed from this?

@pshipton Any other review comments or are you good with this?