@DanHeidinga 2019-11-05
Jenkins test sanity zlinux jdk8