@theresa-m 2019-11-05
> @theresa-m Does this need to be ported to the other extension repos? ie: 11, 13 & jdk?

Yes. I've just updated the description with the 11/13/next pull requests.